Move defopt to package
