+++ lib/defopt/_napoleon.py
